// src/services/api/generated/model/recipeSummary.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Cheffy API
 * A Cheffy API é o backend central de uma aplicação de receitas, projetado para gerenciar catálogo culinário, categorias, tags, ingredientes, informações nutricionais, favoritos de usuário, autenticação via Better Auth, assinatura de upload no Cloudinary e um assistente gastronômico integrado à Groq.
 * OpenAPI spec version: 1.0.0
 */
import type { RecipeSummaryAuthor } from './recipeSummaryAuthor';
import type { RecipeSummaryCategory } from './recipeSummaryCategory';
import type { RecipeSummaryDifficulty } from './recipeSummaryDifficulty';
import type { RecipeSummaryTagsItem } from './recipeSummaryTagsItem';
import type { RecipeSummaryYieldUnit } from './recipeSummaryYieldUnit';

export interface RecipeSummary {
  id: string;
  title: string;
  slug: string;
  description: string;
  /** @nullable */
  imageUrl: string | null;
  prepTime: number;
  cookTime: number;
  totalTime: number;
  yieldAmount: number;
  yieldUnit: RecipeSummaryYieldUnit;
  difficulty: RecipeSummaryDifficulty;
  isPublished: boolean;
  isFeatured: boolean;
  totalFavorites: number;
  views: number;
  deletedAt: string;
  authorId: string;
  categoryId: string;
  isFavorited?: boolean;
  author: RecipeSummaryAuthor;
  category: RecipeSummaryCategory;
  tags: RecipeSummaryTagsItem[];
}